requirements.txt
anyio==4.8.0
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.78.0
fonttools==4.55.3
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.1.0
pydantic==1.10.19
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.19.1
typing_extensions==4.12.2
uvicorn==0.17.6
